refactor: mv navigation model
