--- src/app/modules/shared/models/index.ts
@@ -1,1 +1,0 @@
-export * from './navigationModel';
